Code to execute refunds

+++ handlers/product-move-api/src/main/scala/com/gu/productmove/endpoint/cancel/zuora/GetSubscription.scala
@@ -40,7 +40,7 @@ object GetSubscription {
   case class RatePlanCharge(
     name: String,
     number: String,
-    price: Double,
+    price: BigDecimal,
     billingPeriod: Option[String],
     effectiveStartDate: LocalDate,
     chargedThroughDate: Option[LocalDate],
